creating the base class
and adapting ReachabilityConstraint to it
